requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-ollama
langgraph
numpy
pandas
pyperclip
watchdog

# CUA (Cursor UI Agent) dependencies
pyautogui>=0.9.54
Pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.13.0

# RAG dependencies
sqlite-vec>=0.1.6
pypdf>=6.0.0

# Actuarial dependencies
chainladder>=0.8.21
pydantic>=2.0.0
